Split android_security.c into multiple files
