+++ src/libraries/Native/Unix/System.Security.Cryptography.Native.Android/CMakeLists.txt
@@ -3,9 +3,15 @@ project(System.Security.Cryptography.Native.Android C)
 add_compile_options(-Wno-gnu-zero-variadic-macro-arguments)
 add_compile_options(-Wno-unused-parameter)
 add_compile_options(-Wno-unused-variable)
+add_compile_options(-Wno-unused-function)
 
 set(NATIVECRYPTO_SOURCES
-    android_security.c
+    pal_jni.c
+    pal_misc.c
+    pal_evp.c
+    pal_evp_cipher.c
+    pal_hmac.c
+    pal_bignum.c
 )
 
 add_library(System.Security.Cryptography.Native.Android
