+++ src/libraries/Native/Unix/System.Security.Cryptography.Native.Android/pal_misc.h
@@ -1,0 +1,9 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+
+#pragma once
+
+#include "pal_jni.h"
+
+PALEXPORT int32_t CryptoNative_EnsureOpenSslInitialized(void);
+PALEXPORT int32_t CryptoNative_GetRandomBytes(uint8_t* buf, int32_t num);
